LayoutTests/media/video-source-type-params-expected.txt

EVENT(begin) TEST(relativeURL(video.currentSrc) == 'content/test.mp4') OK
END OF TEST
